Fixes button mash bug, adds 12-button objects
When mashing buttons (intentionally try to break it), a button would not be removed and get stuck in the array. This fix checks for that and resets the array. Also adding back in an enum class that give names to the plethora of hardware buttons I am introducing with the capacitive sense ic. Next is to distribure current functionality across them instead of cramped into two buttons.