adds record to sd card test
will record to sd card whatever is going out audio callback and then save the file on button 10 press. Works pretty flawlessly when logging is enabled. When not, no sound comes through input unless a ubtton like 'play' is pressed. I attribute this to the hacked-in button 10 functionality which was just a test for this commit.