changes sd writing duration to record duration
whatever is recorded to buffer which 'record' button is pressed is now also written (or overwritten) to a new file on the sd card. Additonal WriteProcess() function added because Process() is called from the Audiocallback which does not like to do too much (it freezes the program even to put a printline in the way of crunching those audio bits) so WriteProcess() does all the writing stuff that isn't audiorate dependent and is called from the main for loop. additionally, order of sample variables declaration made to be same as constructor (compiler said to)